+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog of lizard-map
 =======================
 
 
-2.1.10 (unreleased)
+2.1.10 (2011-08-24)
 -------------------
 
 - Requiring lizard-ui 3.0 now, which also means Django 1.3. Updated the